+++ core/mixer/write_frame.h
@@ -54,8 +54,8 @@ public:
 	boost::iterator_range<uint8_t*> image_data(size_t plane_index = 0);	
 	const boost::iterator_range<const uint8_t*> image_data(size_t plane_index = 0) const;
 
-	std::vector<int16_t>& audio_data();
-	const boost::iterator_range<const int16_t*> audio_data() const;
+	std::vector<int32_t>& audio_data();
+	const boost::iterator_range<const int32_t*> audio_data() const;
 	
 	void commit(uint32_t plane_index);
 	void commit();
